moved useful types into their own package
