+++ cmd/reliably/scan/terraform/cmd.go
@@ -1,6 +1,7 @@
 package terraform
 
 import (
+	"github.com/reliablyhq/cli/cmd/reliably/scan/terraform/plan"
 	"github.com/spf13/cobra"
 )
 
@@ -11,6 +12,8 @@ func New() *cobra.Command {
 		Short: "scan terraform resources",
 		Long:  "scan terraform resources for policy violations",
 	}
+
+	cmd.AddCommand(plan.New())
 
 	return cmd
 }
